fix merge error

+++ Box2D/Box2D/Common/b2Math.h
@@ -475,7 +475,7 @@ inline bool operator != (const b2Vec2& a, const b2Vec2& b)
 	return a.x != b.x || a.y != b.y;
 }
 
-inline float32 b2Distance(const b2Vec2& a, const b2Vec2& b)
+inline double b2Distance(const b2Vec2& a, const b2Vec2& b)
 {
 	b2Vec2 c = a - b;
 	return c.Length();
